refactor(onboard): move preset sync out of entrypoint
